pyproject.toml
[project]

name = "shaperglot"
version = "0.4.2"
description = "Test font files for OpenType language support"

license = { file = "README.md" }

authors = [
    { name = "Simon Cozens", email = "simon@simon-cozens.org" }
]

readme = { file = "README.md", content-type = "text/markdown" }

keywords = [
]
classifiers = [
    # TODO: update this list to match your application: https://pypi.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 1 - Planning",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
dependencies= [
"gflanguages >= 0.4.7",
"termcolor >= 1.1.0",
"vharfbuzz",
"strictyaml",
"num2words ~= 0.5",
"ufo2ft",
"youseedee",
"PyYAML",
"protobuf >= 3.19.4",
]

[project.urls]
homepage = "https://pypi.org/project/shaperglot"
documentation = "https://shaperglot.readthedocs.io"
repository = "https://github.com/simoncozens/shaperglot"

[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["Lib"]

[project.scripts]

shaperglot = "shaperglot.cli:main"

[tool.black]

target-version = ["py36", "py37", "py38"]
skip-string-normalization = true

[tool.pylint."messages control"]
good-names = [
    "fh",
    "e",
    "ix",
]
disable=["broad-except", "missing-module-docstring", "no-self-use", "too-many-instance-attributes", "missing-function-docstring", "missing-class-docstring", "duplicate-code","too-many-branches"]
max-line-length=120
